+++ Dockerfile
@@ -1,0 +1,16 @@
+FROM alpine:3.18
+
+
+WORKDIR /app
+
+COPY package*.json ./
+
+RUN npm install
+
+COPY . .
+
+CMD ["npm", "run", "dev"]
+
+
+
+
