+++ MongoDBModels/Post.js
@@ -6,6 +6,7 @@ const PostSchema = new Schema({
     title: String,
     body: String,
     date: { type: Date, default: Date.now },
+    owner_id: String
 });
 
 const PostModel = mongoose.model('Post', PostSchema);
