+++ Validation/authValidationSchemas.js
@@ -1,14 +1,14 @@
 const Joi = require("joi");
 
 const signupSchema = Joi.object({
-  login: Joi.string().required(),
+  username: Joi.string().required(),
   email: Joi.string().email().required(),
   password: Joi.string().required(),
   role: Joi.number().required()
 });
 
 const signinSchema = Joi.object({
-  login: Joi.string().required(),
+  username: Joi.string().required(),
   password: Joi.string().required(),
 });
