+++ controllers/authController.js
@@ -3,10 +3,6 @@ const jwt = require("jsonwebtoken");
 const config = require("../config/config.json");
 const { ApiError } = require("../middleware/ApiError.js");
 const { UserModel, UserUtils } = require("../MongoDBModels/User.js");
-const { error } = require("../Validation/objValidationSchemas.js");
-const fs = require('fs');
-const nodemailer = require('nodemailer');
-const pug = require('pug');
 
 const signup = async (req, res, next) => {
   const user = req.body;
@@ -31,7 +27,7 @@ const signup = async (req, res, next) => {
 const signin = async (req, res, next) => {
   const user = req.body;
 
-  const userDb = await UserUtils.findUserByLogin(user.login);
+  const userDb = await UserUtils.findUserByLogin(user.username);
   console.log(user);
 
   if (!user || !userDb || !bcrypt.compareSync(user.password, userDb.password)) {
@@ -40,13 +36,13 @@ const signin = async (req, res, next) => {
   }
 
   const accessToken = jwt.sign(
-    { id: user.id, login: user.login },
+    { id: userDb._id, username: userDb.username },
     config.env.JWT_SECRET,
     { expiresIn: config.env.tokenExpiration }
   );
 
   const refreshToken = jwt.sign(
-    { id: user.id, login: user.login },
+    { id: user._id, username: user.username },
     config.env.JWT_REFRESH_SECRET
   );
 
@@ -106,113 +102,6 @@ const getMe = async (req, res, next) => {
     .json({ id: user.id, login: user.login, email: user.email });
 };
 
-const newPasswordInit = async (req, res, next) => {
-  try {
-    fs.readFile('./views/init-pass.html', 'utf8', (err, data) => {
-      if (err) {
-        const error = new ApiError(500, "Invalid file reading");
-        return next(error);
-      } else {
-        res.send(data);
-      }
-    });
-  } catch (e) {
-    res.status(500).json({
-      error: e.message,
-    });
-  }
-};
-
-const newPasswordMail = async (req, res, next) => {
-  const email = req.body?.email;
-  if (!email)
-    return next(new ApiError(400, "Can't restore user."));
-  const user = await UserModel.findOne({ email });
-  if (!user)
-    return next(new ApiError(400, "Can't restore user."));
-
-  const newPasswordToken = jwt.sign(
-    { id: user.id, login: user.login },
-    config.env.JWT_NEW_PASSW_SECRET,
-    { expiresIn: config.env.newPasswordTokenExpiration }
-  );
-
-  const link = `http://localhost:3000/reset-password?token=${newPasswordToken}`;
- 
-  const html = pug.renderFile('./views/password-reset-mail.pug', { title: 'Hey', message: 'Hello there!', restorPasswordLink: `${link}` });
-
-  // console.log({ html })
-
-  const transporter = nodemailer.createTransport({
-    host: "smtp.gmail.com",
-    port: 465,
-    secure: true,
-    auth: {
-      user: "gennady.berg@gmail.com",
-      pass: "lnprcjcxddnrqajs",
-    },
-  });
-
-  const info = await transporter.sendMail({
-    from: '', // sender address
-    to: "snack.uventa@gmail.com", // list of receivers
-    subject: "Hello test", // Subject line
-    text: "password : generate new password() ", // plain text body
-    html: html, // html body
-  });
-
-  console.log("Message sent: %s", info.messageId);
-  // Message sent: <d786aa62-4e0a-070a-47ed-0b0666549519@ethereal.email>
-}
-
-const newPasswordTokenEnter = async (req, res, next) => {
-  try {
-    const user = req.user;
-    if (user === error)
-      next(error);
-    else {
-      fs.readFile('./views/create-pass.html', 'utf8', (err, data) => {
-        if (err) {
-          const error = new ApiError(500, "Invalid file reading");
-          return next(error);
-        } else {
-          const newPasswordToken = jwt.sign(
-            { id: user.id, login: user.login },
-            config.env.JWT_NEW_PASSW_SECRET,
-            { expiresIn: config.env.newPasswordTokenExpiration }
-          );
-          data = data.replace("{{token}}", newPasswordToken);
-          res.send(data);
-        }
-      });
-    }
-  } catch (e) {
-    res.status(500).json({
-      error: e.message,
-    });
-  }
-};
-
-const newPasswordTokenSave = async (req, res, next) => {
-  try {
-    const user = req.user;
-    try {
-      const hashedPassword = bcrypt.hashSync(req.password, 10);
-      await UserModel.update({
-        id: user.id,
-        password: hashedPassword,
-      });
-      res.send('password was changed succsessfuly!');
-    } catch (error) {
-      const err = new ApiError(500, "error writing file to db");
-      return next(err);
-    }
-  } catch (e) {
-    res.status(500).json({
-      error: e.message,
-    });
-  }
-};
 
 
 module.exports = {
@@ -220,8 +109,4 @@ module.exports = {
   signin,
   refreshToken,
   getMe,
-  newPasswordTokenEnter,
-  newPasswordTokenSave,
-  newPasswordInit,
-  newPasswordMail
 };
