+++ middleware/resetPasswordStrategy.js
@@ -1,33 +1,48 @@
-const passport = require('passport');
-const LocalStrategy = require('passport-local').Strategy;
-const User = require('../MongoDBModels/User.js'); // Подставьте свой путь к модели пользователя
+const { Strategy } = require('passport-jwt');
+const config = require('../config/config.json');
+const { UserModel } = require('../MongoDBModels/User');
 
-/*passport.use('reset-password', new LocalStrategy(
-  {
-    usernameField: 'email',
-    passwordField: 'password',
-    passReqToCallback: true,
-  },
-  async (req, email, password, done) => {
+let resetPasswEnterOption = {
+  jwtFromRequest: (req) => req.query?.token,
+  secretOrKey: config.env.JWT_NEW_PASSW_SECRET
+};
+
+const resetPasswEnterStrategy =
+  new Strategy(resetPasswEnterOption, async (payload, done) => {
+    const { id } = payload
     try {
-      const user = await User.findOne({ email });
-
-      if (!user) {
-        return done(null, false, { message: 'No user found with that email address.' });
+      user = await UserModel.findById(id)
+      if (user) {
+        done(null, user)
+      } else {
+        done(null, false)
       }
-
-      
-      user.password = password;
-
-      
-      await user.save();
-
-      
-      return done(null, user);
-    } catch (error) {
-      return done(error);
+    } catch (e) {
+      console.log(e)
     }
-  }
-));*/
+  });
 
-module.exports = passport;
+const resetPasswOption = {
+  jwtFromRequest: (req) => req.body?.token,
+  secretOrKey: config.env.JWT_NEW_PASSW_SECRET
+}
+
+const resetPasswStrategy =
+  new Strategy(resetPasswOption, async (payload, done) => {
+    const { id } = payload
+    try {
+      user = await UserModel.findById(id)
+      if (user) {
+        done(null, user)
+      } else {
+        done(null, false)
+      }
+    } catch (e) {
+      console.log(e)
+    }
+  })
+
+module.exports = {
+  resetPasswEnterStrategy,
+  resetPasswStrategy
+};
