+++ package.json
@@ -23,6 +23,7 @@
     "passport": "^0.7.0",
     "passport-jwt": "^4.0.1",
     "passport-local": "^1.0.0",
+    "pm2": "^5.3.1",
     "pug": "^3.0.2",
     "sequelize": "^6.37.1",
     "sqlite": "^5.1.1",
