+++ routes/resetPasswordRouter.js
@@ -1,12 +1,12 @@
 const express = require('express');
-const authController = require('../controllers/authController');
+const resetPasswordController = require('../controllers/resetPasswordController');
 const { passport } = require('../middleware/passport-middleware');
 
 const resetPasswordRouter = express.Router();
 
-resetPasswordRouter.get('/reset-password-init', authController.newPasswordInit);
-resetPasswordRouter.post('/reset-password-mail', authController.newPasswordMail);
-resetPasswordRouter.get('/reset-password', passport.authenticate('reset-password-enter', { session: false }), authController.newPasswordTokenEnter);
-resetPasswordRouter.post('/reset-password', passport.authenticate('reset-password', { session: false }), authController.newPasswordTokenSave);
+resetPasswordRouter.get('/reset-password-init', resetPasswordController.newPasswordInit);
+resetPasswordRouter.post('/reset-password-mail', resetPasswordController.newPasswordMail);
+resetPasswordRouter.get('/reset-password', passport.authenticate('reset-password-enter', { session: false }), resetPasswordController.newPasswordTokenEnter);
+resetPasswordRouter.post('/reset-password', passport.authenticate('reset-password', { session: false }), resetPasswordController.newPasswordTokenSave);
 
 module.exports = resetPasswordRouter;
